[ticket/12039] Do not print success message when config did not exist
PHPBB3-12039

+++ phpBB/phpbb/console/command/config/delete.php
@@ -32,8 +32,15 @@ class delete extends command
 	{
 		$key = $input->getArgument('key');
 
-		$this->config->delete($key);
+		if (isset($this->config[$key]))
+		{
+			$this->config->delete($key);
 
-		$output->writeln("<info>Successfully deleted config $key</info>");
+			$output->writeln("<info>Successfully deleted config $key</info>");
+		}
+		else
+		{
+			$output->writeln("<error>Config $key does not exist</error>");
+		}
 	}
 }
